+++ RELEASE_NOTES.md
@@ -16,7 +16,8 @@ This version improves the layout engine.
 ### ✨ New features
 
 * `BaseKeyboardLayoutProvider` has a new, open `keyboardReturnAction(for:)` function.
-* `KeyboardAction.primary` is a new action type that gathers all primary action types. 
+* `KeyboardAction.primary` is a new action type that gathers all primary action types.
+* `String+sentenceDelimiters` and `wordDelimiters` can now be modified, if you have specific needs. 
 * The iPhone and iPad layout providers can now be injected with locale-specific size providers that can customize the width of certain actions.
 * The iPhone and iPad layout providers now replaces `return` with `done`, `go` and `search` when applicable.